Melchior FRANZ:
After a discussion on irc://irc.flightgear.org/#flightgear and reading
an old thread http://www.google.com/search?q=+site:www.mail-archive.com+%22stall+horn
I think that a patch like the following would be advisable. It turns the
stall horn off on ground if all wheels are on ground and the airspeed is
below 10 kt (arbitrary number). Under these conditions one would assume that
the low pressure conditions for the sensor are not fulfilled. (Rationale:
the whining stall horn on parked/stopped aircrafts is reportedly not realistic
and pretty annoying, too.  :-)